Display trip request and Cost
