+++ src/apiCalls.js
@@ -1,15 +1,10 @@
-import addTripToPage from './domUpdates';
-
 const getData = (suffix) => {
   const baseURL = 'http://localhost:3001/api/v1/';
   return fetch(`${baseURL}${suffix}`).then((response) => response.json());
 };
 
 const getTravelerData = (id) => {
-  // id = `${id}`;
-
   const baseURL = 'http://localhost:3001/api/v1/travelers/';
-
   return fetch(`${baseURL}${id}`).then((response) => response.json());
 };
 
@@ -29,37 +24,3 @@ export const getAllData = (id) => {
     getData('travelers'),
   ]);
 };
-
-// ****************************************Error Handling
-//
-
-// const displayError = (error) => {
-//   if (error) {
-//     document.querySelector('.js-error').innerText =
-//       'You must fill out all fields!';
-//   }
-// };
-
-// const checkForError = (response) => {
-//   console.log(response);
-//   if (!response.ok) {
-//     throw new Error('Please make sure that all fields are filled out.');
-//   } else {
-//     return response.json();
-//   }
-// };
-
-// displayError(response)
-// const displayError = (response) => {
-//   if (!response.ok) {
-//     console.log('Something went wrong, try again!');
-//   } else {
-//     console.log(response.json());
-//     return response.json();
-//   }
-// };
-
-// // .catch((error) => console.error(error));
-// .then((res) => checkForError(res))
-//     .then((trip) => addTripToPage(trip))
-//     .catch((err) => displayError(err));
