+++ src/css/base.scss
@@ -55,17 +55,8 @@ label {
   border: none;
   border-radius: 17px;
   padding-left: 7px;
-  // color: rgb(0, 0, 0);
-  // transition: all 0.25s ease;
 }
 
-
-// #log:hover {
-//   height: 55px;
-//   width: 55px;
-//   padding-bottom: 15px;
-//   border-bottom: solid 1px;
-// }
 
 span  {
   color: white;
@@ -76,10 +67,6 @@ span  {
 float: right;
 background-color: grey;
 }
-
-
-
-
 
 /* body */
